nydia_tools/requirements/requirements-deploy-20240421-add.txt
django-bootstrap4=24.4